added HousePattern. Removed house reference on User
HousePattern is a class with info about the users and appliances but only the aggregated demand pattern. This saves a lot of storage. Also removed the reference to the house in the user class. The house has a list of users, there is no need to store a reference to the house in the user. Last removed the consumption variable in the user. it was not used